split equation_of_motion_utilities.hpp into h and cpp files
